This project helped us to develop our technical skills of python. We were able to learn various techniques to make our project well and good. This project helped to learn about GUI i.e. graphic user interface through which we were able to create a interface through which user can enter values and perform various bitwise operation like or, and, not etc.  We learnt many things through this project to clear all are problems faced during  execution of this project.
This project involves a program on bitwise operation through which we can AND, OR, NOT, XOR operation by using python language. We made two windows in which user will enter values and select which operation to perform and check the result for it. In second window there is the explanation of that operation performed by user. For creation of these window we use GUI and used tkinter library and used various inbuilt and unbuilt function. We used various application of python to implement our code.